packages/hubspot-lead-scoring/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "hubspot-lead-scoring"
version = "0.1.0"
description = "Event-driven lead scoring pipeline for HubSpot contacts"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "requests>=2.31.0",
    "python-dotenv==1.0.1",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scoring*"]
